% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficient_name.R
\name{get_coefficient_name}
\alias{get_coefficient_name}
\title{Creates coefficient variable name by converting to camel case,
adding a prefix and removing symbols}
\usage{
get_coefficient_name(name, prefix = "coef")
}
\arguments{
\item{name}{Raw coefficient name}

\item{prefix}{Prefix before the name of the coefficient. Should be in camel case}
}
\value{
Converted coefficient name
}
\description{
Creates coefficient variable name by converting to camel case,
adding a prefix and removing symbols
}
