% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql.R
\name{toSql}
\alias{toSql}
\alias{toSql,lm-method}
\title{Create SQL function definition from linear model}
\usage{
toSql(
  model,
  fn_title,
  numeric_type = "NUMERIC(24, 6)",
  return_type = "NUMERIC(10,6)",
  ...
)

\S4method{toSql}{lm}(model, fn_title, numeric_type, return_type)
}
\arguments{
\item{model}{Model to create SQL function from}

\item{fn_title}{Name of function to create. Should generally include the schema}

\item{numeric_type}{SQL type to use for numeric parameters. This is also used in the coefficient definitions}

\item{return_type}{SQL type of return value}

\item{...}{Additional arguments}
}
\value{
Response
}
\description{
Generates a user-defined SQL function from a fitted linear model which can then be used for prediction.
}
\examples{
fitted_model <- lm(mpg ~ cyl + wt, data=mtcars)
toSql(fitted_model, "dbo.PredictModel")

}
